Refactoring project layout
