+++ internal/converter/testdata/array_of_enums.go
