+++ internal/converter/testdata/array_of_objects.go
