+++ internal/converter/testdata/array_of_primitives.go
