+++ internal/converter/testdata/nested_object.go
