+++ internal/converter/testdata/proto/ArrayOfEnums.proto
