+++ internal/converter/testdata/proto/ArrayOfObjects.proto
