+++ internal/converter/testdata/proto/Enumception.proto
