+++ internal/converter/testdata/proto/ImportedEnum.proto
