+++ internal/converter/testdata/proto/NestedMessage.proto
