+++ internal/converter/testdata/proto/NestedObject.proto
