+++ internal/converter/testdata/proto/SeveralEnums.proto
