+++ internal/converter/testdata/proto/SeveralMessages.proto
