+++ internal/converter/testdata/second_enum.go
